requirements.txt
python-dotenv == 1.0.0
aiogram == 3.0.0
requests == 2.31.0
multipledispatch == 1.0.0